+++ R/app_ui.R
@@ -98,6 +98,12 @@ app_ui <- function(request) {
         shiny::div(style = nps, page_validationUI("Validation"))
       ),
       bslib::nav_panel(
+        title = "DRMD",
+        icon = shiny::icon("angle-right"),
+        value = "tP_DRDM",
+        shiny::div(style = nps, page_DRMDUI("DRMD"))
+      ),
+      bslib::nav_panel(
         title = "Help",
         icon = shiny::icon("angle-right"),
         value = "tP_help",