+++ NEWS.md
@@ -1,3 +1,14 @@
+# eCerto 0.8.1
+
+* New module DRMD (Digital Reference Material Document)
+* allows to upload a XML file in DRMD format (currently still under development)
+* presents administrative data and measurement data from the XML in 2 tables
+* allows to modify administrative data
+* allows to write modified XML to disc
+* help files for DRMD module initialized
+* report file for DRMD module initialized
+* required 2 new dependencies (`dplyr` and `xml2`)
+
 # eCerto 0.7.4
 
 * Added a `NEWS.md` file to start tracking package changes.